Improve the TestFlxSpriteAlpha.
I've added an animated FlxSprite to check if numFrames, maxFrames and alike are working properly. It's possible to control the animation using the keyboard: F (go to first frame), L (go to last frame), O (go to frame 20, out of bound) and P (play animation).

L and O should produce the same results because the last frame is 15, so FlxSprite.frame = 20 should "auto-fix" to 15 and issue a console warning.